backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
psycopg2-binary==2.9.7
python-decouple==3.8
drf-spectacular==0.27.0
Pillow==10.1.0
django-extensions==3.2.3
factory-boy==3.3.0
pytest-django==4.7.0
pytest==7.4.3
black==23.11.0
flake8==6.1.0
coverage==7.3.2
dj-database-url==2.1.0
